src/requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiogram-forms==1.1.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beanie==1.26.0
certifi==2024.8.30
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.2.2
executing==2.1.0
frozenlist==1.4.1
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
lazy-model==0.2.0
magic-filter==1.0.12
matplotlib-inline==0.1.7
motor==3.5.1
multidict==6.1.0
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
six==1.16.0
stack-data==0.6.3
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
yarl==1.11.1
